src/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==4.0
feedparser==6.0.8
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7